멤버 이름, 필요 없는 부분 삭제 등을 수행

+++ Assets/Scripts/GameSystem/UnitSystem/UnitManager/Handler/UnitManagerDBHandler.cs
@@ -102,13 +102,13 @@ namespace GameSystem.UnitSystem.UnitManager
         public int UnitID;
         public UnitType UnitType;
 
-        public int HP;
+        public float HP;
 
-        public int BaseAttackPower;
+        public float BaseAttackPower;
         public float AttackPowerModifier;
         public float AttackPowerTakenModifier;
 
-        public int BaseHealPower;
+        public float BaseHealPower;
         public float HealPowerModifier;
         public float HealPowerTakenModifier;
     }
